Revert "Add Release notes section to PR body (#296)"
This reverts commit c44a1fa48d217847e30d6df5c5b2afe6a20a5de5.
