+++ src/github/index.ts
@@ -3,4 +3,3 @@ export * from './createGitHub'
 export * from './errors'
 export * from './label'
 export * from './pull-request'
-export * from './releases'
