+++ src/github/pull-request/creator/PullRequestCreator.ts
@@ -6,7 +6,7 @@ import type {
   GitHub,
   Repository as GitHubRepository
 } from '../../GitHub'
-import type { PullRequestBodyCreator } from './PullRequestBodyCreator'
+import { createPullRequestBody } from './createPullRequestBody'
 import type { PullRequestTitleCreator } from './PullRequestTitleCreator'
 
 export class PullRequestCreator {
@@ -14,7 +14,6 @@ export class PullRequestCreator {
   private readonly gitRepo: GitRepository
   private readonly githubRepo: GitHubRepository
   private readonly pullRequestTitleCreator: PullRequestTitleCreator
-  private readonly pullRequestBodyCreator: PullRequestBodyCreator
   private readonly logger: Logger
   private readonly reviewers: string[] | undefined
 
@@ -23,7 +22,6 @@ export class PullRequestCreator {
     gitRepo,
     githubRepo,
     pullRequestTitleCreator,
-    pullRequestBodyCreator,
     logger,
     reviewers
   }: {
@@ -31,7 +29,6 @@ export class PullRequestCreator {
     gitRepo: GitRepository
     githubRepo: GitHubRepository
     pullRequestTitleCreator: PullRequestTitleCreator
-    pullRequestBodyCreator: PullRequestBodyCreator
     logger: Logger
     reviewers?: string[]
   }) {
@@ -39,7 +36,6 @@ export class PullRequestCreator {
     this.gitRepo = gitRepo
     this.githubRepo = githubRepo
     this.pullRequestTitleCreator = pullRequestTitleCreator
-    this.pullRequestBodyCreator = pullRequestBodyCreator
     this.logger = logger
     this.reviewers = reviewers
   }
@@ -54,7 +50,7 @@ export class PullRequestCreator {
     const title = this.pullRequestTitleCreator.create(outdatedPackage)
     this.logger.debug(`title=${title}`)
 
-    const body = await this.pullRequestBodyCreator.create(outdatedPackage)
+    const body = await createPullRequestBody(outdatedPackage)
     this.logger.debug(`body=${body}`)
 
     const pullRequest = await this.github.createPullRequest({
