+++ src/github/pull-request/creator/index.ts
@@ -1,3 +1,2 @@
-export * from './PullRequestBodyCreator'
 export * from './PullRequestCreator'
 export * from './PullRequestTitleCreator'
