--- src/github/releases/fetcher/index.ts
@@ -1,1 +1,0 @@
-export * from './ReleasesFetcher'
