+++ src/main.ts
@@ -14,12 +14,10 @@ import {
   BranchFinder,
   createGitHub,
   LabelCreator,
-  PullRequestBodyCreator,
   PullRequestCloser,
   PullRequestCreator,
   PullRequestFinder,
-  PullRequestTitleCreator,
-  ReleasesFetcher
+  PullRequestTitleCreator
 } from './github'
 import type { Logger } from './logger'
 import { Ncu } from './ncu'
@@ -95,16 +93,11 @@ export const main = async ({
     packageManager: options.packageManager
   })
   const pullRequestTitleCreator = new PullRequestTitleCreator(options.pullRequestTitle)
-  const releasesFetcher = new ReleasesFetcher(github)
-  const pullRequestBodyCreator = new PullRequestBodyCreator({
-    releasesFetcher
-  })
   const pullRequestCreator = new PullRequestCreator({
     github,
     gitRepo,
     githubRepo,
     pullRequestTitleCreator,
-    pullRequestBodyCreator,
     logger,
     reviewers: options.reviewers
   })
